requirements.txt
selenium
psycopg2-binary
flask
webdriver-manager